requirements.txt
streamlit
numpy
Pillow
huggingface_hub
tensorflow==2.13.0
keras==2.13.1
